Codechange: Make non-generic GRF string ids unique per feature.
This prevents engine name strings conflicting with generic strings if more than 53,248 engines are defined.

(cherry picked from commit 0af5d5a9e686bc0e645191b2dc81ff382ee268ec)

# Conflicts:
#	src/newgrf.cpp
#	src/newgrf_text_type.h

+++ src/newgrf_text_type.h
@@ -16,7 +16,7 @@
 #include <vector>
 
 /** Type for GRF-internal string IDs. */
-struct GRFStringIDTag : public StrongType::TypedefTraits<uint16_t, StrongType::Compare, StrongType::Integer> {};
+struct GRFStringIDTag : public StrongType::TypedefTraits<uint32_t, StrongType::Compare, StrongType::Integer> {};
 using GRFStringID = StrongType::Typedef<GRFStringIDTag>;
 
 static constexpr GRFStringID GRFSTR_MISC_GRF_TEXT{0xD000}; ///< Miscellaneous GRF text range.
